migrated the game move to api
- changed the frontend ejs to use axios for sending api requests.
- added gameControlller and injected it in /routes/api/game.router.js and the api router.
- removed an extra semi-colon from the compare-moves test function =))
