+++ routes/api/game.router.js
@@ -1,0 +1,9 @@
+const express = require('express');
+const gameRouter = express.Router();
+
+const HomeController = require('../../app/controllers/gameController');
+
+gameRouter.post("/submitMove", HomeController.submitGameMove);
+
+
+module.exports = gameRouter;
