+++ routes/web/home.router.js
@@ -4,7 +4,6 @@ const homeRouter = express.Router();
 const HomeController = require('../../app/controllers/homeController');
 
 homeRouter.get("/", HomeController.gameHome);
-homeRouter.post("/", HomeController.submitGameMove);
 
 
 module.exports = homeRouter;
