+++ views/home.ejs
@@ -9,21 +9,47 @@
 </head>
 
 <body>
-    <h1>Welcome</h1>
-    <form action="/" method="POST">
-        <label for="moves">Choose a move:</label>
-        <select id="move" name="move">
-            <option value="attack">Attack</option>
-            <option value="block">Block</option>
-            <option value="none">None</option>
-        </select>
-        <input type="submit">
-    </form>
-    <% if (status) { %>
-        <h2>You <%= result %>!</h2>
-    <% } else { %>
-        <%= error %> 
-        <p>Please reflect the page.</p> 
-     <% } %> 
+<h1>Welcome</h1>
+<form class="gameMoveForm">
+    <label for="moves">Choose a move:</label>
+    <select id="move" name="move">
+        <option value="attack">Attack</option>
+        <option value="block">Block</option>
+        <option value="none">None</option>
+    </select>
+    <button>Submit!</button>
+</form>
+<div id="resultOfMatch">
+    <p>You haven't played yet.</p>
+    <p>Please reflect the page.</p>
+</div>
+
+
+
+<script src="https://unpkg.com/axios/dist/axios.min.js"></script>
+<script type="module">
+    document.querySelector('body').addEventListener('click', async event => {
+        event.preventDefault();
+    });
+
+    document.querySelector('form.gameMoveForm button').addEventListener('click', async event => {
+        const target = event.target;
+        const theForm = target.closest('form');
+        const moveValue = theForm.querySelector('#move').value;
+
+        const responseFromApi = await axios.post('/api/games/submitMove',
+            { move: moveValue }
+        );
+
+        const resultDiv = document.querySelector('div#resultOfMatch');
+
+        const result = responseFromApi.data;
+        if (result.status) {
+            resultDiv.innerHTML = `<h2>You ${result.result}!</h2>`;
+        } else {
+            console.log(result);
+        }
+    });
+</script>
 </body>
 </html>
